+++ NEWS.md
@@ -21,6 +21,7 @@
 
 - `model_parameters()` for mixed models gets a `summary_random`-argument to additionally print information about the random effects.
 - `model_parameters()` for *lme*-models now also accepts the `df_method`-argument.
+- The Intercept-parameter in `model_parameters()` for meta-analysis models was renamed to `"Overall"`.
 - `skewness()` gets a `type`-argument, to compute different types of skewness.
 - `kurtosis()` gets a `type`-argument, to compute different types of skewness.